refactor readConfig into utils
